+++ grails-app/services/com/pmf/tihedze/repozitorijradova/services/VolumeService.groovy
@@ -5,6 +5,8 @@ import com.pmf.tihedze.repozitorijradova.Publication
 import com.pmf.tihedze.repozitorijradova.Volume
 import com.pmf.tihedze.repozitorijradova.commands.volume.AddArticlesCommand
 import com.pmf.tihedze.repozitorijradova.commands.volume.CreateVolumeCommand
+import com.pmf.tihedze.repozitorijradova.commands.volume.RemoveArticlesCommand
+import com.pmf.tihedze.repozitorijradova.exceptions.ArticleNotFoundException
 import com.pmf.tihedze.repozitorijradova.exceptions.PublicationNotFoundException
 import com.pmf.tihedze.repozitorijradova.exceptions.VolumeNotFoundException
 import grails.gorm.transactions.Transactional
@@ -12,22 +14,25 @@ import grails.gorm.transactions.Transactional
 @Transactional
 class VolumeService {
 
-    def getAllByPublicationId(UUID publicationId) {
+    List<Volume> getAllByPublicationId(UUID publicationId) {
         final def publication = Publication.findById(publicationId)
         publication.volumes
     }
 
-    def getById(UUID volumeId) {
+    Volume getById(UUID volumeId) {
         Volume.findById(volumeId)
     }
 
-    def create(CreateVolumeCommand command) {
+    List<Volume> getAllByPublicationName(String query) {
+        final List<Publication> publications = Publication.findAllByNameLike(query)
+        publications.collectMany { it.volumes}
+    }
+
+    Volume create(CreateVolumeCommand command) {
         final def publicationUuid = UUID.fromString(command.publicationId)
         final def publication = Publication.findById(publicationUuid)
         final def articlesUuids = command.articleIds.collect {UUID.fromString(it)}
-        final def articles = Article.findAllByIdInList(articlesUuids)
-
-
+        final def articles = Article.getAll(articlesUuids)
 
         if (publication == null) {
             throw new PublicationNotFoundException("No publication found with that id")
@@ -42,13 +47,16 @@ class VolumeService {
         newVolume.save(flush: true)
     }
 
-    def delete(UUID volumeId) {
-        def volume = Volume.findById(volumeId)
+    void delete(UUID volumeId) {
+        final def volume = Volume.findById(volumeId)
+        if (volume  == null) {
+            throw new VolumeNotFoundException("No volume found for the provided id")
+        }
         volume.delete()
     }
 
-    def getByPublicationNameAndIssue(String publicationName, String issue) {
-        final def publication = Publication.findByName(publicationName);
+    Volume getByPublicationNameAndIssue(String publicationName, String issue) {
+        final def publication = Publication.findByName(publicationName)
         if (publication != null) {
             return publication.volumes.find { Volume volume ->
                 volume.issue == issue
@@ -57,19 +65,33 @@ class VolumeService {
         Volume.findByIssue(issue)
     }
 
-    def addArticles(AddArticlesCommand command) {
-        final def articleUuids = command.articleIds.collect {UUID.fromString(it)}
-        final def articlesToAdd = Article.findAllByIdInList(articleUuids)
-        final def volume = Volume.get(UUID.fromString(command.volumeId))
+    Volume addArticles(AddArticlesCommand command, UUID id) {
+        addOrRemoveArticles(command.articleIds, id)
+    }
+
+    Volume removeArticles(RemoveArticlesCommand command, UUID id) {
+        addOrRemoveArticles(command.articleIds, id, true)
+    }
+
+    private static Volume addOrRemoveArticles(List<String> articleIds, UUID volumeId, boolean delete = false) {
+        final def articleUuids = articleIds.collect {UUID.fromString(it)}
+        final def articles = Article.getAll(articleUuids)
+        final def volume = Volume.get(volumeId)
 
         if (volume == null) {
             throw new VolumeNotFoundException('Volume with the provided id not found')
         }
 
-        if (!articlesToAdd.isEmpty()) {
-            volume.addToArticles(articles: articlesToAdd)
+        if (articles.isEmpty()) {
+            throw new ArticleNotFoundException('No articles found for the provided ids')
         }
 
+        if (delete)   {
+            articles.each { volume.removeFromArticles(it)}
+        } else {
+            articles.each { volume.addToArticles(it) }
+        }
 
+        volume.save(flush: true)
     }
 }
